requirements.txt
mcp>=1.0.0
google-generativeai>=0.3.0
boto3>=1.34.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0